requirements.txt
fastapi==0.115.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
uvicorn[standard]==0.30.0
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.13.3
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
bcrypt==4.3.0
python-multipart==0.0.12
itsdangerous==2.2.0
redis==5.1.0
google-cloud-pubsub==2.23.0
jinja2==3.1.4
httpx==0.27.2
loguru==0.7.2
